requirements.txt
tweepy
python-dotenv
textblob
scikit-learn
praw
pandas
numpy
nltk
openpyxl
yfinance
transformers
torch
ta


